Add content and layout for Home section
In this commit, I have added the content and layout for the Home section of the website. This section includes a welcoming message and introduction with the name "Morobang Tshigidimisa," along with a brief description of my passion for data analysis. I have highlighted my expertise in SQL, data visualization, and cloud computing, encouraging visitors to explore my projects. Additionally, I included social media links with icons for LinkedIn, GitHub, Facebook, and Instagram to enhance connectivity. A button for downloading my CV has also been added, along with an image that represents me. These updates create a more engaging and informative introduction to my portfolio.